fix: move  written test to test_document_store and test_document_store_async
